Struct ubip_buffer::BipBufferWriter

pub struct BipBufferWriter<'a, T, St> where
    St: Storage<T>,  { /* fields omitted */ }

The writer handle into a BipBuffer

A write operation is done in two parts:

Prepare a buffer area for writing. A successful prepare returns a mutable slice to the internal buffer.

Commiting of finished data. This allows the data to be written.

A prepared area can be commited in parts or completely in one go, but must be commited from the start of the prepared area.

let mut buffer = [0; 10];
let mut bip_buffer = BipBuffer::new(StorageRef::new(&mut buffer));
let (mut reader, mut writer) = bip_buffer.take_reader_writer().unwrap();
let to_write = writer.prepare(3).unwrap().copy_from_slice(&[1,2,3]);
assert_eq!(reader.values(), []);
writer.commit(1).unwrap();
assert_eq!(reader.values(), [1]);
writer.commit(1).unwrap();
assert_eq!(reader.values(), [1, 2]);
writer.commit(1).unwrap();
assert_eq!(reader.values(), [1, 2, 3]);

Methods

impl<'a, T, St> BipBufferWriter<'a, T, St> where
St: Storage<T>,

pub fn capacity(&self) -> usize

Returns the total capacity of the BipBuffer

This is usually not that useful, depending on where the read pointer is located only parts of the total capacity is availabe.

impl<'a, T, St> BipBufferWriter<'a, T, St> where
St: Storage<T>,

pub fn prepare(&mut self, amount: usize) -> Result<&mut [T], PrepareError>

Try to prepare a set amount of data for commitment.

Depending on the current state of the BipBuffer, data can either be sliced from [write, len), [0, read-1) or [write, read-1).

Arguments

amount - The number of elements requested.

Returns

Success

A slice to the internal buffer with len() equal to amount. This should be written to and commited after the buffer is done.

pub fn prepare_trailing(&mut self) -> Result<&mut [T], PrepareError>

Prepares and returns any part of the buffer trailing the write pointer.

If the layout of the BipBuffer is

|--read---write------len|

then the slice [write, len) will be prepared and returned.

If the layout of the BipBuffer is

|--write----read------len|

then the slice [write, read-1) will be returned (given that read-1-write > 1 holds)

If write == len and read >= 2 then the slice [0, read-1) will be returned.

pub fn prepare_max(&mut self) -> Result<&mut [T], PrepareError>

Prepares and returns the biggest slice of continues data available.

Depending on the state of the buffer this is either located at [write, len), [0, read-1) or [write, read-1)

pub fn commit(&mut self, amount: usize) -> Result<usize, CommitError>

Commits a previsouly prepared part of data.

Commiting data immediately makes the data available for reading.

All data must be commited before new data can be prepared.

Arguments

amount - The number of elements to commit.

pub fn discard(&mut self) -> usize

Discards uncommited but prepared data.

After discard it is safe to prepare a new area for writing.

Returns

The number of elements that were discarded.

Example

let mut buffer = [0; 10];
let mut bip_buffer = BipBuffer::new(StorageRef::new(&mut buffer));
let (mut reader, mut writer) = bip_buffer.take_reader_writer().unwrap();
let to_write = writer.prepare(3).unwrap().copy_from_slice(&[1,2,3]);
assert_eq!(reader.values(), []);
writer.commit(1).unwrap();
assert_eq!(reader.values(), [1]);
assert!(writer.prepare(1).is_err());
assert_eq!(writer.discard(), 2);
assert_eq!(reader.values(), [1]);
let to_write = writer.prepare(2).unwrap();
// This was written but discarded previously
assert_eq!(to_write[0], 2);